move norm to BaseFakeNumpyLinalgNamespace
